Created Routes, authentication and documentation
The routes have been created. however, the directions for email confirmation are still not working.

The user authentication system was created using JWTs. The defined validity for access tokens was 20 minutes, and for refresh tokens 5 days.
Each refresh token can only be used once, after which it will be added to the blacklist.

A documentation was also created, using Swagger, containing the API routes and the necessary parameters (still subject to corrections, if any).